C#: Remove dependency to ASP.NET in the System.Web.cs stub file.
